Move domain::Tags::ExternalBCs to DomainCreator library
