+++ src/Domain/Tags/CMakeLists.txt
@@ -5,7 +5,6 @@ spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
-  ExternalBoundaryConditions.hpp
   FaceNormal.hpp
   Faces.hpp
   SurfaceJacobian.hpp
