+++ tests/Unit/Domain/Tags/CMakeLists.txt
@@ -4,7 +4,6 @@
 set(LIBRARY "Test_DomainTags")
 
 set(LIBRARY_SOURCES
-  Test_ExternalBoundaryConditions.cpp
   Test_Faces.cpp
   Test_SurfaceJacobian.cpp
   )
